+++ languages/c/templates/codeblocks/setter.c
@@ -1,5 +1,5 @@
 /* ${method.rpc.name} - ${method.description} */
-uint32_t ${info.Title}_${method.Name}( ${method.signature.params} )
+int32_t ${info.Title}_${method.Name}( ${method.signature.params} )
 {
     const string method = _T("${info.title.lowercase}.${method.rpc.name}");
 ${if.params}${method.params.serialization}${end.if.params}