+++ languages/c/templates/declarations/polymorphic-pull.c
@@ -1,2 +1,2 @@
 /* ${method.name} - ${method.description} */
-uint32_t ${info.Title}_Push${method.Name}( ${method.signature.params} );
+int32_t ${info.Title}_Push${method.Name}( ${method.signature.params} );