+++ languages/c/templates/declarations/setter.c
@@ -1,0 +1,2 @@
+/* ${method.rpc.name} - ${method.description} */
+int32_t ${info.Title}_${method.Name}( ${method.signature.params} );