+++ languages/c/templates/methods/calls-metrics.c
@@ -2,10 +2,9 @@
 void Metrics_${method.Name}Dispatcher(const void*${if.result} result${end.if.result}) {
     Metrics_${method.Name}(${if.result}(static_cast<${method.result.json.type}>(const_cast<void*>(result)))${end.if.result});
 }
+int32_t ${info.Title}_${method.Name}( ${method.signature.params}${if.result}${if.params}, ${end.if.params}OUT ${method.result.type}* ${method.result.name}${end.if.result}${if.signature.empty}void${end.if.signature.empty} ) {
 
-uint32_t ${info.Title}_${method.Name}( ${method.signature.params}${if.result}${if.params}, ${end.if.params}${method.result.type}* ${method.result.name}${end.if.result}${if.signature.empty}void${end.if.signature.empty} ) {
-
-    uint32_t status = FireboltSDKErrorUnavailable;
+    int32_t status = FireboltSDKErrorUnavailable;
     FireboltSDK::Transport<WPEFramework::Core::JSON::IElement>* transport = FireboltSDK::Accessor::Instance().GetTransport();
     if (transport != nullptr) {
   
@@ -14,7 +13,7 @@ uint32_t ${info.Title}_${method.Name}( ${method.signature.params}${if.result}${i
         status = transport->Invoke("${info.title}.${method.rpc.name}", jsonParameters, jsonResult);
         if (status == FireboltSDKErrorNone) {
             FIREBOLT_LOG_INFO(FireboltSDK::Logger::Category::OpenRPC, FireboltSDK::Logger::Module<FireboltSDK::Accessor>(), "${info.Title}.${method.rpc.name} is successfully invoked");
-${method.result.instantiation}
+    ${method.result.instantiation.with.indent}
 
             void* result = nullptr;
             ${if.result}result = static_cast<void*>(new ${method.result.json.type});${end.if.result}