+++ languages/c/templates/methods/default.c
@@ -1,7 +1,7 @@
 /* ${method.rpc.name} - ${method.description} */
-uint32_t ${info.Title}_${method.Name}( ${method.signature.params}${if.result}${if.params}, ${end.if.params}${method.result.type}* ${method.result.name}${end.if.result}${if.signature.empty}void${end.if.signature.empty} ) {
+int32_t ${info.Title}_${method.Name}( ${method.signature.params}${if.result}${if.params}, ${end.if.params} ${method.result.type}* ${method.result.name}${end.if.result}${if.signature.empty}void${end.if.signature.empty} ) {
 
-    uint32_t status = FireboltSDKErrorUnavailable;
+    int32_t status = FireboltSDKErrorUnavailable;
     FireboltSDK::Transport<WPEFramework::Core::JSON::IElement>* transport = FireboltSDK::Accessor::Instance().GetTransport();
     if (transport != nullptr) {
   
@@ -10,7 +10,7 @@ uint32_t ${info.Title}_${method.Name}( ${method.signature.params}${if.result}${i
         status = transport->Invoke("${info.title.lowercase}.${method.rpc.name}", jsonParameters, jsonResult);
         if (status == FireboltSDKErrorNone) {
             FIREBOLT_LOG_INFO(FireboltSDK::Logger::Category::OpenRPC, FireboltSDK::Logger::Module<FireboltSDK::Accessor>(), "${info.Title}.${method.rpc.name} is successfully invoked");
-${method.result.instantiation}
+    ${method.result.instantiation.with.indent}
         }
   
     } else {