+++ languages/c/templates/methods/property.c
@@ -1,15 +1,13 @@
 /* ${method.rpc.name} - ${method.description} */
-uint32_t ${info.Title}_Get${method.Name}( ${method.signature.params}${if.params}, ${end.if.params}${method.result.type}* ${method.result.name} )
+int32_t ${info.Title}_Get${method.Name}( ${method.signature.params}${if.params}, ${end.if.params} ${method.result.type}* ${method.result.name} )
 {
     const string method = _T("${info.title.lowercase}.${method.rpc.name}");
 ${if.params}${method.params.serialization}${end.if.params}
     ${method.result.json} jsonResult;
-    ${if.params}int status = FireboltSDK::Properties::Get(method, jsonParameters, jsonResult);${end.if.params}
-    ${if.params.empty}int status = FireboltSDK::Properties::Get(method, jsonResult);${end.if.params.empty}
+    ${if.params}int32_t status = FireboltSDK::Properties::Get(method, jsonParameters, jsonResult);${end.if.params}
+    ${if.params.empty}int32_t status = FireboltSDK::Properties::Get(method, jsonResult);${end.if.params.empty}
     if (status == FireboltSDKErrorNone) {
-        if (${method.result.name} != nullptr) {
 ${method.result.instantiation}
-        }
     }
     return status;
 }