+++ languages/c/templates/types/enum.cpp
@@ -1,4 +1,4 @@
     /* ${title} ${description} */
-    ENUM_CONVERSION_BEGIN(${name})
-        { ${NAME}_${key}, _T("${value}") },
-    ENUM_CONVERSION_END(${name})
+    ENUM_CONVERSION_BEGIN(${info.Title}_${name})
+        { ${info.TITLE}_${NAME}_${key}, _T("${value}") },
+    ENUM_CONVERSION_END(${info.Title}_${name})